requirements.txt
aiogram==3.17.0
aiohttp==3.11.11
sqlalchemy==2.0.38
aiosqlite==0.21.0
apscheduler==3.11.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
pytoniq>=0.1.40
